requirements.txt
alpaca-py>=0.28.0
pandas>=2.0.0
ta>=0.10.0
streamlit>=1.28.0
plotly>=5.0.0
pytz>=2023.0
python-dotenv>=1.0.0